Fixed snake

+++ lib/src/typedefs.dart
@@ -731,7 +731,7 @@ typedef sleepDart = void Function(int dwMilliseconds);
 //   DWORD            rop
 // );
 typedef stretchDIBitsNative = Int32 Function(
-    Int64 hdc,
+    IntPtr hdc,
     Int32 xDest,
     Int32 yDest,
     Int32 DestWidth,
